refactor: reorganize concert-related imports and update type references
- Refactored import paths for concert-related functions to improve code organization and clarity.
- Updated type references in various components and utilities to align with the new structure.
- Removed obsolete concert test file to streamline the codebase.
