+++ app/(protected)/layout.tsx
@@ -3,7 +3,7 @@ import { headers } from "next/headers";
 import { redirect } from "next/navigation";
 import Header from "@/components/Header/header";
 import Footer from "@/components/Footer/Footer";
-import { getUserConcertCounts } from "@/lib/concerts";
+import { getUserConcertCounts } from "@/lib/concerts/stats";
 import { checkUserBan } from "@/lib/ban";
 
 export default async function ProtectedLayout({
