+++ app/map/page.tsx
@@ -2,7 +2,8 @@ import React from "react"
 import { redirect } from "next/navigation"
 import Layout from "../../src/components/layout-client"
 import MapClient from "../../src/components/MapClient"
-import { getUserConcerts, getUserConcertCounts } from "@/lib/concerts"
+import { getUserConcerts } from "@/lib/concerts/read"
+import { getUserConcertCounts } from "@/lib/concerts/stats"
 import { auth } from "@/lib/auth"
 import { headers } from "next/headers"
 import type { Metadata } from "next"
