+++ src/lib/__tests__/concerts/delete.test.ts
@@ -1,0 +1,29 @@
+import { describe, test, expect, beforeEach, vi } from "vitest"
+import { prisma } from "@/lib/prisma"
+import { deleteConcert } from "@/lib/concerts/mutations/delete"
+
+describe("deleteConcert", () => {
+  beforeEach(() => {
+    vi.clearAllMocks()
+  })
+
+  test("test_deleteConcert_non_attendee_returns_false", async () => {
+    vi.mocked(prisma.userConcert.findUnique).mockResolvedValue(null)
+    const result = await deleteConcert("concert-1", "user-2")
+    expect(result).toBe(false)
+  })
+
+  test("test_deleteConcert_removes_attendance_and_orphaned_concert", async () => {
+    vi.mocked(prisma.userConcert.findUnique).mockResolvedValue({
+      id: "attendance-1",
+      userId: "user-1",
+      concertId: "concert-1",
+    } as any)
+    vi.mocked(prisma.userConcert.delete).mockResolvedValue({} as any)
+    vi.mocked(prisma.userConcert.count).mockResolvedValue(0)
+    vi.mocked(prisma.concert.delete).mockResolvedValue({} as any)
+    const result = await deleteConcert("concert-1", "user-1")
+    expect(result).toBe(true)
+    expect(prisma.concert.delete).toHaveBeenCalledWith({ where: { id: "concert-1" } })
+  })
+})
