+++ src/lib/__tests__/concertsPublicProfileRedaction.contract.test.ts
@@ -1,0 +1,8 @@
+import { describe, test, expect } from "vitest"
+
+describe("concerts public profile redaction contract", () => {
+  test("test_contract_file_exists_for_public_profile_redaction_suite", () => {
+    expect(true).toBe(true)
+  })
+})
+
