+++ src/lib/concerts/errors.ts
@@ -1,0 +1,8 @@
+/** Thrown when the user already has this concert in their list (matched by date/location/headliner). */
+export class ConcertAlreadyExistsError extends Error {
+  constructor(public concertId: string) {
+    super("Concert already in list")
+    this.name = "ConcertAlreadyExistsError"
+  }
+}
+
